+++ stylelib/ops.h
@@ -45,7 +45,7 @@ public:
     static QPalette::ColorRole opposingRole(const QPalette::ColorRole &role);
     static QPalette::ColorRole bgRole(const QWidget *w, const QPalette::ColorRole fallBack = QPalette::Window);
     static QPalette::ColorRole fgRole(const QWidget *w, const QPalette::ColorRole fallBack = QPalette::WindowText);
-    static void toolButtonData(const QToolButton *tbtn, bool &nextsel, bool &prevsel, bool &isintop, unsigned int &sides);
+    static void toolButtonData(const QToolButton *tbtn, bool &nextsel, bool &prevsel, bool &isintop, quint8 &sides);
     static bool hasMenu(const QToolButton *tb, const QStyleOptionToolButton *stb = 0);
     static void swap(int &t1, int &t2);
     template<typename T>static void swap(T &t1, T &t2)
